SA-CORE-2025-003 by shin24, anzuukino, mcdruid, nicxvan, ghost of drupal past, longwave

+++ core/modules/views/src/DisplayPluginCollection.php
@@ -6,6 +6,7 @@ use Drupal\Component\Plugin\Exception\PluginException;
 use Drupal\Component\Plugin\PluginManagerInterface;
 use Drupal\Core\Plugin\DefaultLazyPluginCollection;
 use Drupal\Core\StringTranslation\StringTranslationTrait;
+use Drupal\views\Plugin\views\display\DisplayPluginInterface;
 
 /**
  * A class which wraps the displays of a view so you can lazy-initialize them.
@@ -63,7 +64,9 @@ class DisplayPluginCollection extends DefaultLazyPluginCollection {
    */
   public function clear() {
     foreach (array_filter($this->pluginInstances) as $display) {
-      $display->destroy();
+      if ($display instanceof DisplayPluginInterface) {
+        $display->destroy();
+      }
     }
 
     parent::clear();
